+++ parser/src/defs.rs
@@ -1263,6 +1263,15 @@ primitive!(
     /// ex: ⊂₃ 1 2_3_4 5_6
     /// ex: ⊂₄ "al" "ig" "at" "or"
     ///
+    /// Sided subscripted [join] uses the argument on that side as the list, automatically promoting the rank if necessary.
+    /// ex: ⊂⌟ 1_2 3_4
+    ///   : ⊂⌞ 1_2 3_4
+    /// ex: ⊂⌟ 1 2_3
+    ///   : ⊂⌞ 1_2 3
+    /// [un] sided [join] unjoins from that side.
+    /// ex: °⊂⌟ [1 2 3 4]
+    ///   : °⊂⌞ [1 2 3 4]
+    ///
     /// [join]ing to the front of an array is a bit slower than [join]ing to the back because it requires all the existing rows to be shifted.
     ///
     /// [join]'s glyph is `⊂` because it kind of looks like a magnet pulling its two arguments together.